requirements.txt
faiss_cpu==1.7.4
fastapi==0.121.3
PyMuPDF==1.23.8
langchain==1.0.8
numpy==2.3.5
openai==2.8.1
prometheus_client==0.19.0
pydantic==2.12.4
python-dotenv==1.2.1
rank_bm25==0.2.2
redis==4.6.0
Requests==2.32.5
sentence_transformers==5.1.0
torch==2.8.0
uvicorn==0.38.0
